Refactor search index processing to dedicated type
